+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.1.1
+### Fixes
+- Added missing crafting recipes for Silver Bulb and its variants.
+- Added missing smelting recipes for Silver Ore and Deepslate Silver Ore blocks
+
+
 # 1.1.0
 ### Additions
 - Added Silver Bulb (and all its oxidized/waxed variants)